demos/shortify/requirements.txt
aiohttp==3.11.18
aiohttp-jinja2==1.6
pyyaml==6.0.2
redis==6.1.0
trafaret==2.1.1
pytest-aiohttp==1.0.4
pytest-asyncio==0.23.5
